feat: 流水线取消当次构建并启动新构建 #5930
