+++ src/backend/ci/core/openapi/biz-openapi/src/main/kotlin/com/tencent/devops/openapi/resources/apigw/v3/ApigwBuildResourceV3Impl.kt
@@ -244,6 +244,16 @@ class ApigwBuildResourceV3Impl @Autowired constructor(
         )
     }
 
+    override fun buildRefresh(userId: String, projectId: String, pipelineId: String, buildId: String): Result<String> {
+        logger.info("buildRefresh $userId|$projectId|$pipelineId|$buildId")
+        return client.get(ServiceBuildResource::class).buildRefresh(
+            userId = userId,
+            projectId = projectId,
+            pipelineId = pipelineId,
+            buildId = buildId
+        )
+    }
+
     companion object {
         private val logger = LoggerFactory.getLogger(ApigwBuildResourceV3Impl::class.java)
     }
