+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/api/ServiceBuildResourceImpl.kt
@@ -564,6 +564,15 @@ class ServiceBuildResourceImpl @Autowired constructor(
         )
     }
 
+    override fun buildRefresh(userId: String, projectId: String, pipelineId: String, buildId: String): Result<String> {
+        return Result(pipelineBuildFacadeService.refreshBuild(
+            userId = userId,
+            projectId = projectId,
+            pipelineId = pipelineId,
+            buildId = buildId
+        ))
+    }
+
     private fun checkParam(projectId: String, pipelineId: String) {
         if (pipelineId.isBlank()) {
             throw ParamBlankException("Invalid pipelineId")
